requirements.txt
streamlit>=1.35.0
numpy>=1.26.0
groq>=0.9.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
pypdf>=4.0.0
pdfplumber>=0.11.0
rank-bm25>=0.2.2
langchain-text-splitters>=0.2.0
